fix extract-translate
