+++ package.json
@@ -55,7 +55,7 @@
     "@storybook/react": "5.1.8",
     "@types/deep-freeze": "^0.1.1",
     "@types/flat": "0.0.28",
-    "@types/jest": "24.0.14",
+    "@types/jest": "24.0.15",
     "@types/lodash.flattendeep": "^4.4.6",
     "@types/lodash.pick": "4.4.6",
     "@types/node": "12.0.8",
@@ -87,7 +87,7 @@
     "typescript": "3.5.2"
   },
   "dependencies": {
-    "antd": "3.19.3",
+    "antd": "3.19.4",
     "axios": "0.19.0",
     "connected-react-router": "6.4.0",
     "deep-freeze": "0.0.1",
@@ -99,7 +99,7 @@
     "lodash.pick": "4.4.0",
     "react": "16.8.6",
     "react-dom": "16.8.6",
-    "react-i18next": "10.11.0",
+    "react-i18next": "10.11.1",
     "react-redux": "7.1.0",
     "react-responsive": "7.0.0",
     "react-router-dom": "5.0.1",
