+++ src/resources/locales/index.ts
