--- src/resources/translations/en/translation.json
@@ -1,1 +1,0 @@
-{}
